Check for existence of .pinned files before trying to back them up.

+++ jenkins-backup.sh
@@ -27,7 +27,9 @@ mkdir -p "$ARC_DIR/"{plugins,jobs,users}
 
 cp "$JENKINS_HOME/"*.xml "$ARC_DIR"
 cp "$JENKINS_HOME/plugins/"*.[hj]pi "$ARC_DIR/plugins"
-cp "$JENKINS_HOME/plugins/"*.[hj]pi.pinned "$ARC_DIR/plugins"
+if [ -f "$JENKINS_HOME/plugins/"*.[hj]pi.pinned ] ; then
+  cp "$JENKINS_HOME/plugins/"*.[hj]pi.pinned "$ARC_DIR/plugins"
+fi
 cp -R "$JENKINS_HOME/users/"* "$ARC_DIR/users"
 
 cd "$JENKINS_HOME/jobs/"
